pyproject.toml
[project]
name = "ai-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=3.1.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.0",
]

[project.scripts]
ai-agents = "ai_agents.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
